Add missing rust-src SHAs.
This adds missing known SHAs for the rust-src download so it can be cached
in the repository cache and used in an airgapped environment.

Additionally, this fixes a bug in the `load_rust_src` macro that was not
using the correct key to lookup the known SHA for the `rust-src` tarball
download.

This will enable the use of rust-analyzer in an airgapped environment.

Signed-off-by: Timothy Trippel <ttrippel@google.com>